fix compiler warning

+++ src/net/SerialProto.cpp
@@ -11,9 +11,10 @@ bool SerialProto::send(const pb_msgdesc_t *fields, const void *src_struct)
 
 bool SerialProto::receive(const pb_msgdesc_t *fields, void *dst_struct)
 {
-    pb_istream_t stream = {SerialProto::decode_callback, &_serial, _serial.available(), 0};
+    pb_istream_t stream = {SerialProto::decode_callback, &_serial, (size_t)_serial.available(), 0};
     return pb_decode(&stream, fields, dst_struct);
 }
+
 
 bool SerialProto::encode_callback(pb_ostream_t *stream, const uint8_t *buf, size_t count)
 {
